requirements.txt
# Python 3.10+ required (google-genai supports 3.10–3.14). Tested on 3.12.7.
flask>=2.2.0,<3.0.0
google-genai>=0.3.0
pydantic>=2.5.0
requests>=2.31.0
python-dotenv>=0.21.0,<1.0.0
streamlit>=1.20.0,<2.0.0
pytest>=7.0.0

